+++ packages/nimble-components/src/select/tests/select.spec.ts
@@ -482,8 +482,35 @@ describe('Select', () => {
             element,
             'registerOption'
         ).and.callThrough();
-        registerOptionSpy.calls.reset();
         element.insertBefore(newOption, element.options[0]!);
+
+        expect(registerOptionSpy.calls.count()).toBe(1);
+        expect(element.options).toContain(newOption);
+
+        // While the option is registered synchronously as shown above,
+        // properties like selectedIndex will only be correct asynchronously
+        // See https://github.com/ni/nimble/issues/1915
+        expect(element.selectedIndex).toBe(0);
+        await waitForUpdatesAsync();
+        expect(element.value).toBe('one');
+        // This assertion shows that after 'slottedOptionsChanged' runs, the
+        // 'selectedIndex' state has been corrected to expected DOM order.
+        expect(element.selectedIndex).toBe(1);
+        await disconnect();
+    });
+
+    it('option added directly to DOM under group synchronously registers with Select', async () => {
+        const { element, connect, disconnect } = await setupWithGroups();
+        await connect();
+        await waitForUpdatesAsync();
+        const newOption = new ListOption('foo', 'foo');
+        const registerOptionSpy = spyOn(
+            element,
+            'registerOption'
+        ).and.callThrough();
+        const pageObject = new SelectPageObject(element);
+        const group = pageObject.getAllGroups()[0]!;
+        group.insertAdjacentElement('afterbegin', newOption);
 
         expect(registerOptionSpy.calls.count()).toBe(1);
         expect(element.options).toContain(newOption);
@@ -2000,10 +2027,16 @@ describe('Select', () => {
             ]);
         });
 
-        it('exercise getOptionLabels', async () => {
+        it('exercise getGroupOptionLabelsByIndex', async () => {
             await clickAndWaitForOpen(element);
-            const optionLabels = pageObject.getGroupOptionLabels(0);
+            const optionLabels = pageObject.getGroupOptionLabelsByIndex(0);
             expect(optionLabels).toEqual(['One', 'Two', 'Edge']);
+        });
+
+        it('exercise getGroupOptionLabelsByLabel', async () => {
+            await clickAndWaitForOpen(element);
+            const optionLabels = pageObject.getGroupOptionLabelsByLabel('Group Two');
+            expect(optionLabels).toEqual(['Three', 'Four']);
         });
     });
 });